feat: update validation for refresh method

+++ app/Services/JWTServiceImpl.php
@@ -36,7 +36,7 @@ class JWTServiceImpl implements JWTService
      */
     public function refreshTokens(string $token, JWTSubject $user): array
     {
-        if ($this->getRefreshTokenByToken($token)) {
+        if ($this->getRefreshTokenByToken($token) && $this->isRefreshExpired($token)) {
             $this->refreshTokenRepository->delete($token);
             return $this->generateTokens($user);
         } else {
